requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pymongo==4.6.3
mysql-connector-python==8.3.0
qdrant-client[httpx]==1.7.0
sentence-transformers==2.6.1
python-dotenv==1.0.1
